pyproject.toml
[project]
name = "autoresetlol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "customtkinter>=5.2.2",
    "keyboard>=0.13.5",
    "packaging>=25.0",
    "pillow>=12.0.0",
    "psutil>=7.1.3",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.17.0",
    "pyscreeze>=1.0.1",
    "pywin32>=311",
]
